+++ src/Altinn.Apps/AppTemplates/AspNet/Altinn.App.Common/Models/Components.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+
+namespace Altinn.App.Common.Models
+{
+    /// <summary>
+    /// UI Components
+    /// </summary>
+    public class Components
+    {
+        /// <summary>
+        /// Exclude from pdf
+        /// </summary>
+        public List<string> ExcludeFromPdf { get; set; }
+    }
+}
