+++ src/Altinn.Apps/AppTemplates/AspNet/Altinn.App.Common/Models/LayoutSets.cs
@@ -1,0 +1,15 @@
+using System.Collections.Generic;
+
+namespace Altinn.App.Common.Models
+{
+    /// <summary>
+    /// Layoutset
+    /// </summary>
+    public class LayoutSets
+    {
+        /// <summary>
+        /// Sets
+        /// </summary>
+        public List<LayoutSet> Sets { get; set; }
+    }
+}
